+++ toolbox/+igutils/mustBePartition.m
@@ -1,10 +1,19 @@
 function mustBePartition(membership)
 %MUSTBEPARTITION assert the argument is a valid membership vector
 
-    eid = "igraph:NotMembershipVector";
+    eid = "igraph:invalidVertixVector";
     msg = "";
-    if ~isnumeric(membership)
+
+    if ~isnumeric(membership) && ~islogical(membership)
         msg = "Membership vector must be numeric or logical";
+    end
+
+    if min(membership) < 1
+        msg = "Node IDs must start at 1";
+    end
+
+    if ~allfinite(membership) || ~all(membership == floor(membership), 'all')
+        msg = "All nodes IDs must be integers"
     end
 
     if ~isreal(membership)